Move GuildPermissionLevel enum to shared subproject
