+++ Quarterdeck/src/main/java/com/fredboat/backend/quarterdeck/db/repositories/impl/SqlSauceGuildPermsRepo.java
@@ -27,7 +27,7 @@ package com.fredboat.backend.quarterdeck.db.repositories.impl;
 
 import com.fredboat.backend.quarterdeck.db.entities.main.GuildPermissions;
 import com.fredboat.backend.quarterdeck.db.repositories.api.GuildPermsRepo;
-import com.fredboat.backend.quarterdeck.rest.v1.transfer.GuildPermissionLevel;
+import com.fredboat.backend.shared.GuildPermissionLevel;
 import org.springframework.stereotype.Component;
 import space.npstr.sqlsauce.DatabaseWrapper;
 import space.npstr.sqlsauce.entities.SaucedEntity;
